requirements.txt
requests
schedule
python-dotenv
beautifulsoup4
psycopg2-binary
